Tree rows paint all the way across
Trees paint the selection across the whole row and clicking the mouse
anywhere on the row effectively selects the row. Required some
extremely fiddly workarounds for the icon in Aqua look and feel not
having a variant for the selected row and because of the way the
workaround works, animation doesn’t occur if the row is selected.